+++ sdk/resourcehealth/azure-resourcemanager-resourcehealth/src/main/java/com/azure/resourcemanager/resourcehealth/ResourceHealthManager.java
@@ -10,11 +10,13 @@ import com.azure.core.http.HttpPipeline;
 import com.azure.core.http.HttpPipelineBuilder;
 import com.azure.core.http.HttpPipelinePosition;
 import com.azure.core.http.policy.AddDatePolicy;
+import com.azure.core.http.policy.AddHeadersFromContextPolicy;
 import com.azure.core.http.policy.HttpLogOptions;
 import com.azure.core.http.policy.HttpLoggingPolicy;
 import com.azure.core.http.policy.HttpPipelinePolicy;
 import com.azure.core.http.policy.HttpPolicyProviders;
 import com.azure.core.http.policy.RequestIdPolicy;
+import com.azure.core.http.policy.RetryOptions;
 import com.azure.core.http.policy.RetryPolicy;
 import com.azure.core.http.policy.UserAgentPolicy;
 import com.azure.core.management.http.policy.ArmChallengeAuthenticationPolicy;
@@ -23,9 +25,15 @@ import com.azure.core.util.Configuration;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.resourcehealth.fluent.MicrosoftResourceHealth;
 import com.azure.resourcemanager.resourcehealth.implementation.AvailabilityStatusesImpl;
+import com.azure.resourcemanager.resourcehealth.implementation.EventOperationsImpl;
+import com.azure.resourcemanager.resourcehealth.implementation.EventsOperationsImpl;
+import com.azure.resourcemanager.resourcehealth.implementation.ImpactedResourcesImpl;
 import com.azure.resourcemanager.resourcehealth.implementation.MicrosoftResourceHealthBuilder;
 import com.azure.resourcemanager.resourcehealth.implementation.OperationsImpl;
 import com.azure.resourcemanager.resourcehealth.models.AvailabilityStatuses;
+import com.azure.resourcemanager.resourcehealth.models.EventOperations;
+import com.azure.resourcemanager.resourcehealth.models.EventsOperations;
+import com.azure.resourcemanager.resourcehealth.models.ImpactedResources;
 import com.azure.resourcemanager.resourcehealth.models.Operations;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
@@ -39,6 +47,12 @@ public final class ResourceHealthManager {
     private AvailabilityStatuses availabilityStatuses;
 
     private Operations operations;
+
+    private ImpactedResources impactedResources;
+
+    private EventsOperations eventsOperations;
+
+    private EventOperations eventOperations;
 
     private final MicrosoftResourceHealth clientObject;
 
@@ -68,6 +82,19 @@ public final class ResourceHealthManager {
     }
 
     /**
+     * Creates an instance of ResourceHealth service API entry point.
+     *
+     * @param httpPipeline the {@link HttpPipeline} configured with Azure authentication credential.
+     * @param profile the Azure profile for client.
+     * @return the ResourceHealth service API instance.
+     */
+    public static ResourceHealthManager authenticate(HttpPipeline httpPipeline, AzureProfile profile) {
+        Objects.requireNonNull(httpPipeline, "'httpPipeline' cannot be null.");
+        Objects.requireNonNull(profile, "'profile' cannot be null.");
+        return new ResourceHealthManager(httpPipeline, profile, null);
+    }
+
+    /**
      * Gets a Configurable instance that can be used to create ResourceHealthManager with optional configuration.
      *
      * @return the Configurable instance allowing configurations.
@@ -85,6 +112,7 @@ public final class ResourceHealthManager {
         private final List<HttpPipelinePolicy> policies = new ArrayList<>();
         private final List<String> scopes = new ArrayList<>();
         private RetryPolicy retryPolicy;
+        private RetryOptions retryOptions;
         private Duration defaultPollInterval;
 
         private Configurable() {
@@ -146,6 +174,19 @@ public final class ResourceHealthManager {
         }
 
         /**
+         * Sets the retry options for the HTTP pipeline retry policy.
+         *
+         * <p>This setting has no effect, if retry policy is set via {@link #withRetryPolicy(RetryPolicy)}.
+         *
+         * @param retryOptions the retry options for the HTTP pipeline retry policy.
+         * @return the configurable object itself.
+         */
+        public Configurable withRetryOptions(RetryOptions retryOptions) {
+            this.retryOptions = Objects.requireNonNull(retryOptions, "'retryOptions' cannot be null.");
+            return this;
+        }
+
+        /**
          * Sets the default poll interval, used when service does not provide "Retry-After" header.
          *
          * @param defaultPollInterval the default poll interval.
@@ -178,7 +219,7 @@ public final class ResourceHealthManager {
                 .append("-")
                 .append("com.azure.resourcemanager.resourcehealth")
                 .append("/")
-                .append("1.0.0-beta.2");
+                .append("1.0.0");
             if (!Configuration.getGlobalConfiguration().get("AZURE_TELEMETRY_DISABLED", false)) {
                 userAgentBuilder
                     .append(" (")
@@ -196,10 +237,15 @@ public final class ResourceHealthManager {
                 scopes.add(profile.getEnvironment().getManagementEndpoint() + "/.default");
             }
             if (retryPolicy == null) {
-                retryPolicy = new RetryPolicy("Retry-After", ChronoUnit.SECONDS);
+                if (retryOptions != null) {
+                    retryPolicy = new RetryPolicy(retryOptions);
+                } else {
+                    retryPolicy = new RetryPolicy("Retry-After", ChronoUnit.SECONDS);
+                }
             }
             List<HttpPipelinePolicy> policies = new ArrayList<>();
             policies.add(new UserAgentPolicy(userAgentBuilder.toString()));
+            policies.add(new AddHeadersFromContextPolicy());
             policies.add(new RequestIdPolicy());
             policies
                 .addAll(
@@ -230,7 +276,11 @@ public final class ResourceHealthManager {
         }
     }
 
-    /** @return Resource collection API of AvailabilityStatuses. */
+    /**
+     * Gets the resource collection API of AvailabilityStatuses.
+     *
+     * @return Resource collection API of AvailabilityStatuses.
+     */
     public AvailabilityStatuses availabilityStatuses() {
         if (this.availabilityStatuses == null) {
             this.availabilityStatuses = new AvailabilityStatusesImpl(clientObject.getAvailabilityStatuses(), this);
@@ -238,12 +288,52 @@ public final class ResourceHealthManager {
         return availabilityStatuses;
     }
 
-    /** @return Resource collection API of Operations. */
+    /**
+     * Gets the resource collection API of Operations.
+     *
+     * @return Resource collection API of Operations.
+     */
     public Operations operations() {
         if (this.operations == null) {
             this.operations = new OperationsImpl(clientObject.getOperations(), this);
         }
         return operations;
+    }
+
+    /**
+     * Gets the resource collection API of ImpactedResources.
+     *
+     * @return Resource collection API of ImpactedResources.
+     */
+    public ImpactedResources impactedResources() {
+        if (this.impactedResources == null) {
+            this.impactedResources = new ImpactedResourcesImpl(clientObject.getImpactedResources(), this);
+        }
+        return impactedResources;
+    }
+
+    /**
+     * Gets the resource collection API of EventsOperations.
+     *
+     * @return Resource collection API of EventsOperations.
+     */
+    public EventsOperations eventsOperations() {
+        if (this.eventsOperations == null) {
+            this.eventsOperations = new EventsOperationsImpl(clientObject.getEventsOperations(), this);
+        }
+        return eventsOperations;
+    }
+
+    /**
+     * Gets the resource collection API of EventOperations.
+     *
+     * @return Resource collection API of EventOperations.
+     */
+    public EventOperations eventOperations() {
+        if (this.eventOperations == null) {
+            this.eventOperations = new EventOperationsImpl(clientObject.getEventOperations(), this);
+        }
+        return eventOperations;
     }
 
     /**