+++ sdk/resourcehealth/azure-resourcemanager-resourcehealth/src/main/java/com/azure/resourcemanager/resourcehealth/fluent/MicrosoftResourceHealth.java
@@ -10,8 +10,7 @@ import java.time.Duration;
 /** The interface for MicrosoftResourceHealth class. */
 public interface MicrosoftResourceHealth {
     /**
-     * Gets Subscription credentials which uniquely identify Microsoft Azure subscription. The subscription ID forms
-     * part of the URI for every service call.
+     * Gets The ID of the target subscription.
      *
      * @return the subscriptionId value.
      */
@@ -58,4 +57,25 @@ public interface MicrosoftResourceHealth {
      * @return the OperationsClient object.
      */
     OperationsClient getOperations();
+
+    /**
+     * Gets the ImpactedResourcesClient object to access its operations.
+     *
+     * @return the ImpactedResourcesClient object.
+     */
+    ImpactedResourcesClient getImpactedResources();
+
+    /**
+     * Gets the EventsOperationsClient object to access its operations.
+     *
+     * @return the EventsOperationsClient object.
+     */
+    EventsOperationsClient getEventsOperations();
+
+    /**
+     * Gets the EventOperationsClient object to access its operations.
+     *
+     * @return the EventOperationsClient object.
+     */
+    EventOperationsClient getEventOperations();
 }